+++ Glyssen/Controls/Browser.cs
@@ -3,7 +3,6 @@ using System.ComponentModel;
 using System.ComponentModel.Design;
 using System.Windows.Forms;
 using Gecko;
-using Gecko.DOM;
 using Gecko.Events;
 using SIL.Windows.Forms.HtmlBrowser;
 
@@ -55,10 +54,7 @@ namespace Glyssen.Controls
 
 		public void ScrollElementIntoView(string elementId, int adjustment = 0)
 		{
-			var element = m_geckoBrowser.Document.GetElementById(elementId);
-			if (element == null)
-				return;
-			var div = new GeckoDivElement(element.DomObject);
+			var div = m_geckoBrowser.Document.GetElementById(elementId) as GeckoHtmlElement;
 			div.ScrollIntoView(true);
 			div.Parent.ScrollTop += adjustment;
 		}