+++ Glyssen/Program.cs
@@ -4,8 +4,10 @@ using System.Configuration;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
+using System.Reflection;
 using System.Windows.Forms;
 using DesktopAnalytics;
+using Gecko;
 using Glyssen.Dialogs;
 using Glyssen.Properties;
 using Glyssen.Shared;
@@ -19,9 +21,11 @@ using PtxUtils;
 using SIL;
 using SIL.IO;
 using SIL.Reporting;
+using SIL.Windows.Forms.FileSystem;
 using SIL.Windows.Forms.i18n;
 using SIL.Windows.Forms.Reporting;
 using SIL.WritingSystems;
+using Analytics = GlyssenEngine.Utilities.Analytics;
 
 namespace Glyssen
 {
@@ -54,6 +58,8 @@ namespace Glyssen
 			}
 
 			MessageModal.Default = new WinFormsMessageBox();
+			Analytics.Default = new WinFormsAnalytics();
+			Fonts.Default = new WinFormsFonts();
 
 			Application.EnableVisualStyles();
 			Application.SetCompatibleTextRenderingDefault(false);
@@ -139,13 +145,13 @@ namespace Glyssen
 			//}
 
 #if DEBUG
-			using (new Analytics("jBh7Qg4jw2nRFE8j8EY1FDipzin3RFIP", userInfo))
+			using (new DesktopAnalytics.Analytics("jBh7Qg4jw2nRFE8j8EY1FDipzin3RFIP", userInfo))
 #else
 			//default is to allow tracking if this isn't set
 			string feedbackSetting = Environment.GetEnvironmentVariable("FEEDBACK")?.ToLower();
 			var allowTracking = string.IsNullOrEmpty(feedbackSetting) || feedbackSetting == "yes" || feedbackSetting == "true";
 
-			using (new Analytics("WEyYj2BOnZAP9kplKmo2BDPvfyofbMZy", userInfo, allowTracking))
+			using (new DesktopAnalytics.Analytics("WEyYj2BOnZAP9kplKmo2BDPvfyofbMZy", userInfo, allowTracking))
 #endif
 			{
 				foreach (var exception in _pendingExceptionsToReportToAnalytics)
@@ -184,6 +190,8 @@ namespace Glyssen
 				if (!sldrIsInitialized)
 					Sldr.Initialize();
 
+				Xpcom.Initialize(XulRunnerLocation);
+
 				try
 				{
 					Application.Run(new MainForm(args));
@@ -208,12 +216,78 @@ namespace Glyssen
 			}
 		}
 
+		private static string XulRunnerLocation
+		{
+			get
+			{
+				// Firefox files should exist in the "Firefox64" subfolder of the folder containing the executable.
+				string firefoxPath = Path.Combine(Application.StartupPath, "Firefox64");
+				if (File.Exists(Path.Combine(firefoxPath, "xul.dll")))
+					return firefoxPath;
+
+				// But while running in development - look for the nuget packages location.
+				var version = FileVersionInfo.GetVersionInfo(Assembly.GetAssembly(typeof(GeckoWebBrowser)).Location).ProductVersion;
+				string codeBase = Assembly.GetExecutingAssembly().CodeBase;
+				UriBuilder uri = new UriBuilder(codeBase);
+				string path = Uri.UnescapeDataString(uri.Path);
+
+				string fullPath = SearchForFirefoxDlls(path, version);
+				if (fullPath == null)
+					fullPath = SearchForFirefoxDlls(path, version + "-alpha");
+				if (fullPath == null)
+					fullPath = SearchForFirefoxDlls(path, version.TrimEnd('0').TrimEnd('.') + "-alpha");
+				if (fullPath != null)
+					return fullPath;
+#if DEBUG
+				// Search back from the source file. This is to allow the form designer to initalize geckofx based controls.
+				fullPath = SearchForFirefoxDlls(new FileInfo(__FILE__()).DirectoryName, version);
+				if (fullPath != null)
+				{
+					return fullPath;
+				}
+				throw new ApplicationException($"Unable to locate XulRunner files for Firefox FirefoxHtmlEditor. {codeBase} + {Environment.CurrentDirectory} + {(__FILE__())}");
+#else
+                throw new ApplicationException($"Unable to locate XulRunner files for Firefox FirefoxHtmlEditor. {codeBase} + {Environment.CurrentDirectory}");
+#endif
+			}
+		}
+
+		private static string SearchForFirefoxDlls(string path, string version)
+		{
+			string fullPath;
+			var dir = new DirectoryInfo(Path.GetDirectoryName(path));
+			do
+			{
+				string bits = Environment.Is64BitProcess ? "64." : "32.";
+				fullPath = Path.Combine(dir.FullName,
+					Path.Combine("packages", "Geckofx60." + bits + version, "content", "Firefox"));
+				if (Directory.Exists(fullPath))
+					return fullPath;
+				fullPath = Path.Combine(dir.FullName,
+					Path.Combine("packages", "Geckofx60." + bits + version.TrimEnd('0').TrimEnd('.'), "content",
+						"Firefox"));
+				if (Directory.Exists(fullPath))
+					return fullPath;
+				dir = dir.Parent;
+			}
+			while (dir != null);
+
+			return null;
+		}
+
+#if DEBUG
+		static string __FILE__([System.Runtime.CompilerServices.CallerFilePath] string fileName = "")
+		{
+			return fileName;
+		}
+#endif
+
 		private static void HandleDeleteUserSettings(string userConfigSettingsPath)
 		{
 				var confirmationString = Localizer.GetString("Program.ConfirmDeleteUserSettingsFile",
 					"Do you want to delete your user settings? (This will clear your most-recently-used project, publishing settings, UI language settings, etc.  It will not affect your project data.)");
 
-			if (MessageResult.Yes == MessageModal.Show(confirmationString, GlyssenInfo.kProduct, Buttons.YesNo, Icon.Warning))
+				if (DialogResult.Yes == MessageBox.Show(confirmationString, GlyssenInfo.kProduct, MessageBoxButtons.YesNo, MessageBoxIcon.Warning))
 					File.Delete(userConfigSettingsPath);
 		}