+++ Glyssen/ReferenceText.cs
@@ -3,13 +3,11 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
-using Glyssen.Character;
 using Glyssen.Shared;
 using Glyssen.Shared.Bundle;
 using GlyssenEngine.Character;
 using SIL.Reporting;
 using SIL.Scripture;
-using SIL.Xml;
 
 namespace Glyssen
 {
@@ -249,8 +247,8 @@ namespace Glyssen
 			} : (Action < PortionScript >)null;
 
 			var matchup = new BlockMatchup(vernacularBook, iBlock, splitBlocks,
-			nextVerse => IsOkayToSplitAtVerse(nextVerse, vernacularBook.Versification, verseSplitLocationsBasedOnRef),
-			this, predeterminedBlockCount);
+				nextVerse => IsOkayToSplitAtVerse(nextVerse, vernacularBook.Versification, verseSplitLocationsBasedOnRef),
+				this, predeterminedBlockCount);
 
 			if (!matchup.AllScriptureBlocksMatch)
 			{