+++ GlyssenEngine/Character/ControlCharacterVerseData.cs
@@ -2,16 +2,12 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using Glyssen.Properties;
 using Glyssen.Shared;
-using GlyssenEngine.Character;
-using L10NSharp.TMXUtils;
-using L10NSharp.UI;
 using SIL.Extensions;
 using SIL.Scripture;
 using static System.String;
 
-namespace Glyssen.Character
+namespace GlyssenEngine.Character
 {
 	public class ControlCharacterVerseData : CharacterVerseData
 	{
@@ -44,12 +40,10 @@ namespace Glyssen.Character
 			if (TabDelimitedCharacterVerseData == null)
 				TabDelimitedCharacterVerseData = Resources.CharacterVerseData;
 			LoadAll();
-			LocalizeItemDlg<TMXDocument>.StringsLocalized += HandleStringsLocalized; // Don't need to unsubscribe since this object will be around as long as the program is running.
 		}
 
 		private void Dispose()
 		{
-			LocalizeItemDlg<TMXDocument>.StringsLocalized -= HandleStringsLocalized;
 			s_singleton = null;
 		}