+++ GlyssenEngine/Character/ICharacter.cs
@@ -1,0 +1,9 @@
+﻿namespace GlyssenEngine.Character
+{
+	public interface ICharacter
+	{
+		bool IsNarrator { get; }
+		string LocalizedDisplay { get; }
+		string CharacterId { get; }
+	}
+}